+++ package.json
@@ -17,24 +17,27 @@
     "build": "rollup -c",
     "lint": "tsc --noEmit",
     "test": "tsc -v",
-    "prepublish": "yarn build"
+    "prepublishOnly": "yarn build"
   },
   "dependencies": {
     "@infomaximum/assert": "^1.0.0",
     "lodash.get": "^4.4.2"
   },
+  "peerDependencies": {},
   "devDependencies": {
     "@infomaximum/config": "^2.0.0",
-    "@rollup/plugin-commonjs": "24.0.1",
-    "@rollup/plugin-node-resolve": "15.0.1",
-    "@rollup/plugin-typescript": "11.0.0",
-    "@types/lodash.get": "^4.4.7",
-    "rollup": "3.20.0",
-    "rollup-plugin-copy": "3.4.0",
+    "@rollup/plugin-commonjs": "28.0.3",
+    "@rollup/plugin-node-resolve": "16.0.1",
+    "@rollup/plugin-typescript": "12.1.2",
+    "@types/lodash.get": "^4.4.9",
+    "prettier": "3.5.3",
+    "rollup": "4.41.1",
+    "rollup-plugin-copy": "3.5.0",
+    "rollup-plugin-delete": "^3.0.1",
+    "rollup-plugin-dts": "^6.2.1",
     "standard-version": "9.5.0",
-    "prettier": "2.8.6",
     "tslib": "^2.4.1",
-    "typescript": "5.0.2"
+    "typescript": "5.8.3"
   },
   "directories": {
     "lib": "lib"
