--- rollup.config.cjs
@@ -1,36 +1,0 @@
-const packageJSON = require("./package.json");
-const typescript = require("@rollup/plugin-typescript");
-const resolve = require("@rollup/plugin-node-resolve");
-const commonjs = require("@rollup/plugin-commonjs");
-
-const externalPackages = [
-  ...Object.keys(packageJSON.dependencies || {}),
-  ...Object.keys(packageJSON.peerDependencies || {}),
-];
-
-const regexesOfPackages = externalPackages.map(
-  (packageName) => new RegExp(`^${packageName}(\/.*)?`)
-);
-
-/** @type {import("rollup").RollupOptions[]} */
-const config = [
-  {
-    input: "index.ts",
-    output: [
-      {
-        file: "dist/index.js",
-        format: "cjs",
-        sourcemap: true,
-      },
-      {
-        file: "dist/index.esm.js",
-        format: "es",
-        sourcemap: true,
-      },
-    ],
-    plugins: [typescript(), resolve(), commonjs()],
-    external: regexesOfPackages,
-  },
-];
-
-module.exports = config;
